# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mi API – Práctica 5",
    "version": "1.0.0",
    "description": "Documentación automática con OpenAPI/Swagger"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "Estado del servicio"
    },
    {
      "name": "Auth",
      "description": "Autenticación y JWT"
    },
    {
      "name": "Users",
      "description": "Operaciones con usuarios"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a3f4e7c2-5af8-4b2b-a1b2-9c9f12345678"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@dominio.com"
          },
          "name": {
            "type": "string",
            "example": "Ada Lovelace"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ]
      },
      "UserInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "nuevo@dominio.com"
          },
          "name": {
            "type": "string",
            "example": "Nueva Persona"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "Secreta123!"
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@dominio.com"
          },
          "password": {
            "type": "string",
            "example": "Secreta123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Detalle del error"
          }
        },
        "required": [
          "message"
        ]
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Inicia sesión y devuelve un JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "examples": {
                "ejemploValido": {
                  "value": {
                    "email": "user@dominio.com",
                    "password": "Secreta123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login correcto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                },
                "examples": {
                  "ok": {
                    "value": {
                      "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Revisa el estado del servicio",
        "description": "Retorna { ok: true } para indicar que el servicio responde.",
        "responses": {
          "200": {
            "description": "Servicio operativo",
            "content": {
              "application/json": {
                "example": {
                  "ok": true
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Lista paginada de usuarios (requiere auth)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Número de página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Tamaño de página"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 10
                    }
                  }
                },
                "examples": {
                  "ejemplo": {
                    "value": {
                      "data": [
                        {
                          "id": "a3f4e7c2-5af8-4b2b-a1b2-9c9f12345678",
                          "email": "user@dominio.com",
                          "name": "Ada Lovelace"
                        }
                      ],
                      "page": 1,
                      "limit": 10
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Crea un usuario (requiere auth)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obtiene un usuario por id (requiere auth)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "UUID del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    }
  }
}